+++ apps/api-journeys/src/modules/application/index.ts
@@ -1,8 +1,6 @@
 import { createApplication } from 'graphql-modules'
 import { schemaBuilder } from '@core/shared/util-graphql'
-import { blockModule } from '../block'
-import { journeyModule } from '../journey'
-import { responseModule } from '../response'
+import { blockModule, journeyModule, responseModule } from '..'
 
 export default createApplication({
   modules: [blockModule, journeyModule, responseModule],
