+++ apps/api-journeys/src/modules/index.ts
@@ -1,0 +1,5 @@
+export { actionModule } from './action'
+export { blockModule } from './block'
+export { journeyModule } from './journey'
+export { responseModule } from './response'
+export { signUpModule } from './signUp'
