+++ apps/api-journeys/src/modules/response/transformResponse/index.ts
@@ -1,2 +1,1 @@
 export { transformResponse } from './transformResponse'
-export { responseModule } from './response'
