+++ apps/api-journeys/src/modules/signUp/index.ts
@@ -1,0 +1,1 @@
+export { signUpModule } from './signUp'
